+++ spec/resque-history/plugins/history_spec.rb
@@ -18,8 +18,8 @@ class MaxHistoryJob
   end
 end
 
-
 describe Resque::Plugins::History do
+
   it "should be compliance with Resqu::Plugin document" do
     expect { Resque::Plugin.lint(Resque::Plugins::History) }.to_not raise_error
   end
@@ -34,7 +34,7 @@ describe Resque::Plugins::History do
       arr = Resque.redis.lrange(Resque::Plugins::History::HISTORY_SET_NAME, 0, -1)
 
       arr.count.should == 1
-      JSON.parse(arr.first).should == {"class"=>"HistoryJob", "args"=>[12], "time"=>"2000-09-01 12:00"}
+      JSON.parse(arr.first).should == {"class"=>"HistoryJob", "args"=>[12], "time"=>"2000-09-01 12:00", "execution"=>0}
     end
   end
 
@@ -70,9 +70,9 @@ describe Resque::Plugins::History do
 
       arr.count.should == 3
       arr.should == [
-          {"class"=>"HistoryJob", "args"=>[11], "time"=>"2000-09-01 12:00"},
-          {"class"=>"HistoryJob", "args"=>[12], "time"=>"2000-09-01 12:00"},
-          {"class"=>"HistoryJob", "args"=>[13], "time"=>"2000-09-01 12:00"}
+          {"class"=>"HistoryJob", "args"=>[11], "time"=>"2000-09-01 12:00", "execution"=>0},
+          {"class"=>"HistoryJob", "args"=>[12], "time"=>"2000-09-01 12:00", "execution"=>0},
+          {"class"=>"HistoryJob", "args"=>[13], "time"=>"2000-09-01 12:00", "execution"=>0}
       ].collect(&:to_json)
     end
 
@@ -101,4 +101,122 @@ describe Resque::Plugins::History do
 
   end
 
+
+  describe "record execution time" do
+
+    it "should record execution time" do
+
+      Resque.enqueue(SlowHistoryJob, 10)
+
+      Timecop.freeze(Time.local(2000, 9, 1, 12, 0, 0)) do
+        job = Resque.reserve('test')
+        job.perform
+      end
+
+      arr = Resque.redis.lrange(Resque::Plugins::History::HISTORY_SET_NAME, 0, -1)
+
+      arr.count.should == 1
+      arr.should == [
+          {"class"=>"SlowHistoryJob", "args"=>[10], "time"=>"2000-09-01 12:10", "execution"=>600}
+      ].collect(&:to_json)
+
+    end
+
+    it "should not confuse different job times" do
+
+      Resque.enqueue(SlowHistoryJob, 10)
+      Resque.enqueue(SlowHistoryJob, 5)
+
+      Timecop.freeze(Time.local(2000, 9, 1, 12, 0, 0)) do
+        job = Resque.reserve('test')
+        job.perform
+
+        job = Resque.reserve('test')
+        job.perform
+
+      end
+
+      arr = Resque.redis.lrange(Resque::Plugins::History::HISTORY_SET_NAME, 0, -1)
+
+      arr.count.should == 2
+      arr.should == [
+          {"class"=>"SlowHistoryJob", "args"=>[5], "time"=>"2000-09-01 12:15", "execution"=>300},
+          {"class"=>"SlowHistoryJob", "args"=>[10], "time"=>"2000-09-01 12:10", "execution"=>600}
+      ].collect(&:to_json)
+
+    end
+
+    it "should record times in the order of completion not in order of starting" do
+
+      #start the longer job first but it should finish last
+      Resque.enqueue(SleepyHistoryJob, 3)
+      Resque.enqueue(SleepyHistoryJob, 1)
+
+      job = Resque.reserve('test')
+      job.perform
+
+      job = Resque.reserve('test')
+      job.perform
+
+      arr = Resque.redis.lrange(Resque::Plugins::History::HISTORY_SET_NAME, 0, -1)
+
+
+      JSON.parse(arr.first)["class"].should =="SleepyHistoryJob"
+      JSON.parse(arr.first)["args"].should ==[1]
+      JSON.parse(arr.first)["execution"].should ==1
+
+
+      JSON.parse(arr.last)["class"].should =="SleepyHistoryJob"
+      JSON.parse(arr.last)["args"].should ==[3]
+      JSON.parse(arr.last)["execution"].should ==3
+
+    end
+
+    it "should record failed jobs" do
+
+      Resque.enqueue(ExceptionJob,"nothing")
+
+      job = Resque.reserve('test')
+
+      lambda { job.perform }.should raise_exception
+      arr = Resque.redis.lrange(Resque::Plugins::History::HISTORY_SET_NAME, 0, -1)
+
+
+      arr.count.should == 1
+
+      JSON.parse(arr.first)["class"].should =="ExceptionJob"
+      JSON.parse(arr.first)["args"].should == ["nothing"]
+      JSON.parse(arr.first)["execution"].should == nil
+      JSON.parse(arr.first)["error"].should == "I'm an error"
+    end
+
+  end
+
+end
+
+class SlowHistoryJob
+  extend Resque::Plugins::History
+  @queue = :test
+
+  def self.perform(time_in_minutes)
+    Timecop.travel(Time.now + 60*time_in_minutes)
+  end
+end
+
+class SleepyHistoryJob
+  extend Resque::Plugins::History
+  @queue = :test
+
+  def self.perform(time_in_seconds)
+    sleep(time_in_seconds)
+  end
+end
+
+class ExceptionJob
+  extend Resque::Plugins::History
+  @queue = :test
+
+  def self.perform(arg)
+    raise StandardError, "I'm an error"
+  end
 end